Updating abaqus solid in-house meshing example and tools, renaming shell directory and shell.py to mesh_gen
