+++ docs/user-guide/solid_models.rst
@@ -52,6 +52,8 @@ An example called `cubit_solid.py` exists in the examples folder.
 Discontinuous meshes in Abaqus
 ------------------------------
 
+The example pyNuMAD/examples/write_abaqus_solid_model.py generates a fully solid element model input file in Abaqus format, which can be imported and analyzed using Abaqus CAE.  With the ih-house mesher, the shear webs and adhesive are tied to the blade outer shell with nodal constraint equations.  These are included in the model input file generated by this example.
+
 Layered solid model (homogenized sandwich panels)
 ==================================================
