+++ src/pynumad/shell/surface.py
@@ -1,6 +1,6 @@
 import numpy as np
-import pynumad.shell.mesh_tools as mt
-from pynumad.shell.shell_region import *
+import pynumad.mesh_gen.mesh_tools as mt
+from pynumad.mesh_gen.shell_region import *
 
 
 class Surface:
